move smart tests
