+++ lib/modules/smart/test/fixtures/dynamic_registration.json
@@ -1,0 +1,1 @@
+{"client_id":"72e731c6-648c-4947-af23-4518a312c845","client_id_issued_at":1518316591,"registration_access_token":"eyJraWQiOiJyc2ExIiwiYWxnIjoiUlMyNTYifQ.eyJhdWQiOiI3MmU3MzFjNi02NDhjLTQ5NDctYWYyMy00NTE4YTMxMmM4NDUiLCJpc3MiOiJodHRwczpcL1wvc2ItYXV0aC5zbWFydGhlYWx0aGl0Lm9yZ1wvIiwiaWF0IjoxNTE4MzE2NTkxLCJqdGkiOiI3M2E3MmYyMS1lN2EzLTQyYWYtODc2Ni0yYTM4NWE5NGU2MGUifQ.E3rAYVXltDOkRyLJjUxHct_8AM7O-paxKpLZ3hEkUCXkHfec6zCrnfYYbUfAguJ645Ahl2_avMbI9YxP7oqfYSO_hgxHEoE5MXA_mnp7wjJvwQp_DKlh4OrOvFV8gxyusXTVoP3IBu7VN_kqZKmoV-4QzxaKN9_GqU0YN8SPbgZRgLoawmgLXlwmhjblEufUFjt5mPcQoFsx-hQ15HX3kzb2egPMGNAlv_DMpuHxxIUjRC5NY1_Hz9xI7JlVD4PHetq35m3pc-hleunb45mPH5Bk7gjwShXIPcHHDslWZawKFU3Ho3EGJpB9eGZgvhQ4gSntLvMsJUFJAVbPqxrkgw","registration_client_uri":"https://sb-auth.smarthealthit.org/register/72e731c6-648c-4947-af23-4518a312c845","redirect_uris":["http://localhost:4567/smart/4ffksQ/3mWi7k/redirect"],"client_name":"Inferno","token_endpoint_auth_method":"none","scope":"launch/patient online_access openid profile launch user/*.* patient/*.*","grant_types":["authorization_code"],"response_types":["code"],"initiate_login_uri":"http://localhost:4567/smart/4ffksQ /3mWi7k/launch"}
