+++ lib/modules/smart/test/fixtures/openid_configuration.json
@@ -1,0 +1,14 @@
+{
+	"registration_endpoint": "https://www.example.com/register",
+	"token_endpoint": "https://www.example.com/token",
+	"token_endpoint_auth_methods_supported": ["client_secret_post", "client_secret_basic", "none"],
+	"jwks_uri": "https://www.example.com/jwk",
+	"id_token_signing_alg_values_supported": ["HS256", "HS384", "HS512", "RS256", "RS384", "RS512", "none"],
+	"authorization_endpoint": "https://www.example.com/authorize",
+	"introspection_endpoint": "https://www.example.com/introspect",
+	"response_types_supported": ["code"],
+	"grant_types_supported": ["authorization_code"],
+	"scopes_supported": ["launch", "openid", "patient/*.*", "profile"],
+	"userinfo_endpoint": "https://www.example.com/userinfo",
+	"issuer": "https://www.example.com/"
+}
