+++ lib/modules/smart/test/fixtures/standalone_token_exchange.json
@@ -1,0 +1,8 @@
+{
+    "access_token":"eyJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIwYWYwNmU5My0zMTJhLTQwZGEtOTlkYS00ZDVjYjUyYzYwZTQiLCJpc3MiOiJodHRwczpcL1wvc2ItYXV0aC5zbWFydGhlYWx0aGl0Lm9yZ1wvIiwiZXhwIjoxNTE4Mzc5MjgzLCJpYXQiOjE1MTgzNzU2ODMsImp0aSI6IjNmMWUwYmJmLWQ2MDMtNGE1Ni04MDEzLTZkYjYxYzVkMWVhNyJ9.iVnDGZD3WgFZu7BMTwNUoZ--xHedHfoeE9EfisGDNsn5qmh8AJfcyyM55v8BgqfQ7szMqdZctOPl4xfJoi5OxQT3SpMfYQ8vPQg1T530YZt5ik6OUP86QpvE0VEEM4cEq1YDHyDl3FsU4oKyb6uJR7PwULiG6Qpij0ydzQdhLrQZpV_hP7_2pyLDnNVT9R7xtCDoo0pxsh4MmCFjrbmgaXohLCl1aqIgCx58LSZnARH_WXYF9345aS63irc_DmyMXlBQbPq31HxHjH80HHajoPs-3oMWZSLuNfDwUZo2GHasLvZYzdWA7gkaolRRLbZDPx9TYXwzfEJ1YwYYulkYFw",
+    "token_type":"Bearer",
+    "expires_in":3599,
+    "scope":"launch/patient online_access openid profile launch user/*.* patient/*.*",
+    "patient":"783a1f02-5bf9-41c9-90d0-c2c4d35d3ec3",
+    "id_token":"eyJraWQiOiJyc2ExIiwiYWxnIjoiUlMyNTYifQ.eyJzdWIiOiJyc2NhbmxvbkBtaXRyZS5vcmciLCJhdWQiOiIwYWYwNmU5My0zMTJhLTQwZGEtOTlkYS00ZDVjYjUyYzYwZTQiLCJvcmdhbml6YXRpb25OYW1lIjoiOVJ0bzRxd1IwbXNSbEpXbjBpOGxsQll1TEg0ZnZvdE0iLCJkaXNwbGF5TmFtZSI6IlJvYmVydCBTY2FubG9uIiwicHJvZmlsZSI6InN5NjF5dlJnemtxMDZPNUQxbmYwIiwia2lkIjoicnNhMSIsImlzcyI6Imh0dHBzOlwvXC9zYi1hdXRoLnNtYXJ0aGVhbHRoaXQub3JnXC8iLCJleHAiOjE1MTgzNzYyODMsImlhdCI6MTUxODM3NTY4M30.WZxm11AXLguFX8IUPLC_mozfad0s0i_1um09cOfPQ4-GpUSMIFFMx7xt5_qFX5b-lCZKcHC7BBdHdp5Qnbf_azSCjBh9BHG08FcI9Q8OorftP2O5RefTXY1WVHaEbYJcRokBUfwOQYk3PqXmfyVdruxxqMZwF0cCP-EsOSpxGEcPgGqV2zr266vsS2iDUXDVmS1kLZqEPaHqT_v2OkOwZKUy8Gpz1LxY_g48m5RKGblYZjYUIQUBaLov3aDIQQRy6gdtGAyhMbWehe5q8FNsdZvUMOP2761SIieq39Lv0H-54uJnEDWXJJkyWJ17dw3USCqe2X0yRRxflnpLTNXARQ"
+}
