+++ lib/modules/smart/test/standalone_launch_test.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative '../test_helper'
+require_relative '../../../../test/test_helper'
 
 class StandaloneLaunchSequenceTest < MiniTest::Test
   def setup
